fixed linting, project structure
