+++ src/components/route-editor/RouteEditor.jsx
@@ -1,18 +1,43 @@
 import React from "react";
+import MediaQuery from "react-responsive";
+import classNames from "classnames";
 import BackButton from "../common/BackButton";
-import MapContainer from "../../containers/route-editor/Map";
-// import MapEditorComponent from "./MapEditor";
+import MapForm from "../../containers/route-editor/Form";
+import MapComponent from "../../containers/route-editor/Map";
 
-const RouteEditor = () => {
-  return (
-    <div className={"route-editor"}>
-      <BackButton />
-      <div className={"route-editor__content"}>
-        {/* <MapEditor /> */}
-        <MapContainer />
+export default class RouteEditor extends React.Component {
+
+  state = { isRouteListOpen: false };
+
+  toggleMapForm = (e) => {
+    e.preventDefault();
+    const { isRouteListOpen } = this.state;
+    this.setState({ isRouteListOpen: !isRouteListOpen });
+  }
+
+  render() {
+    const { isRouteListOpen } = this.state;
+
+    return (
+      <div className={"route-editor"}>
+        <BackButton />
+        <div className={"route-editor__content"}>
+          <MediaQuery maxWidth={768}>
+            <button
+              className={classNames("button", isRouteListOpen && "button--close")}
+              onClick={this.toggleMapForm}
+            >
+              {isRouteListOpen ? "Закрыть" : "Маршрут"}
+            </button>
+            {isRouteListOpen && (<MapForm />)}
+          </MediaQuery>
+          <MediaQuery minWidth={769}>
+            <MapForm />
+          </MediaQuery>
+          <MapComponent />
+        </div>
       </div>
-    </div>
-  );
-};
+    );
+  }
 
-export default RouteEditor;
+}
