+++ src/components/route-editor/form/PointItem.jsx
@@ -3,7 +3,10 @@ import PropTypes from "prop-types";
 import { SortableElement } from "react-sortable-hoc";
 
 const PointItem = (props) => {
-  const { point, onDeleteButtonClick } = props;
+  const {
+    point,
+    onDeleteButtonClick
+  } = props;
 
   return (
     <li className={"map-form__list-item"}>
@@ -16,9 +19,9 @@ const PointItem = (props) => {
   );
 };
 
-// PointItem.propTypes = {
-//   point: PropTypes.object,
-//   onDeleteButtonClick: PropTypes.func
-// };
+PointItem.propTypes = {
+  point: PropTypes.object,
+  onDeleteButtonClick: PropTypes.func
+};
 
 export default SortableElement(PointItem);
