+++ src/lib/config.js
@@ -1,5 +1,3 @@
-const config = {
-  mode: process.env.NODE_ENV
-};
+const config = { mode: process.env.NODE_ENV };
 
 export default config;
