+++ src/scss/base/_base.scss
@@ -2,3 +2,4 @@
 @import "./fonts";
 @import "./variables";
 @import "./common";
+@import "./GUI";
