+++ src/scss/base/common.scss
@@ -46,7 +46,6 @@ img {
 	min-width: 320px;
 	max-width: 1440px;
 	height: 100%;
-	// min-height: 622px;
 	margin: 0 auto;
 }
